scripts/usb_capture/requirements.txt
# Requirements for USB capture scripts
# Install with: pip install -r requirements.txt

pyusb>=1.2.1
